Function prototypes changed, changed variables to correct ones
Renamed variables to match in all files.
Function prototypes changed to reflect the change in usage of variables
Yk is the bin count of each subinterval, not the random variable that represents it (?). The random variables array is no longer used - Y stays constant for the duration of the run, only the lambda values are changed. xk represents the midpoint of the kth interval, so changed to reflect this. Not sure if things are correct.

Still does not work for a number of subintervals different to the interval time.