Fixed interval value calculations, IWLS b recalculation error fixed
Calculations of start, midpoint and end of intervals were not being correctly shifted to account for the possibility in start time shifts.
When b does not fall within the required constraints in IWLS, a factor of 2 was being applied incorrectly. This seems to point to an error in Massey et al 1996, p 369, "easy to see how estimators are related when we make adjustments to satisfy the nonnegativity constraints ..."