Less file reading during estimation, rewrite complete
New function to extract events within a specific interval from an array of events. This should mean that the event data has to be read only once.
Rewrite of the piecewise estimator is complete. The estimation function is shorter and more readable, and the line extension function has also been improved.