Automake appears to compile stuff. Not tested yet.
Note that src/Makefile.am contains the location of the muParser library files, because for some reason cannot find another way to do it (too time consuming)
This just adds -I/usr/include/muParser to the compilation of the generator. Ideally this would be done automatically, but am stumped so far.
aclocal.m4 contains some macros that get used. Not really sure if this is how things should be done, so adding just in case.