Launcher started. Removed mains from libraries.
Param file has more parameters to specify estimator behaviour
Generation now uses the launcher to start
Estimator wrapper which is used to run specific estimators
Added -no-install to launcher to makefile to make it so that it does not get written as a script, but as an executable. Allows running of valgrind on it.