Generation of gaussians more structured, more math functions
Testing and implementation of math functions to find min value in array, add constant to all values in an array, and generate an array with the same values in it.
Can call the launcher with -s to generate gaussians. Currently outputs to files.
Default parameters added for gaussian generation, added to paramfile generation.