Delta estimator area method working.
More parameters for the delta estimator.
Some functions moved from combinefunction to general util because they're useful in other places too.
Launcher passes parameters to the delta estimator from the paramfile.
Delta estimator area method uses a single function to do the work, with some helpers. Due to different types of return from the gaussian and baseline estimators it takes a void pointer and the type of estimate, and it will change the calculation based on that. This stops code duplication.