PMF delay estimator improved, helper functions for various things
Functions in math_util for calculating pmf sums, getting gaussian sums at many points, finding the largest absolute value in an array.
PMF delay estimator prototype has a better structure and is almost working.
Function combination modified. Should now receive an array containing n-1 value, where n is the number of functions being combined. Previously the first function had a delay as well, but this one is considered to be zero-based and therefore have no delay. Not sure if this change will stay, because it causes some issues with the generator.
Functions for initialising structs.