Progress on pmf estimator. Underlying logic working reasonably well.
Various fixes to the launcher and combinefunctions functions to work correctly with the per-stream delay system.
pmf estimator almost working. Need to work out how best to get values of the pmf when the combined function does not have a delay for all points in the interval. Try using more subintervals.