Experimenter executes experiments, dynamic parameters, estimator specification
Experiment execution happens, but not yet in a useful way.
Experiments can be added to the parameter file and executed by running the experimenter, so long as certain parameters are provided.
Estimator to use for each experiment can be specified.
Suppressed most print statements in estimators with the VERBOSE variable.