More experiment output granularity, parameter output
Experiment data is now output to separate directories for each experiment, along with the parameter settings used for that experiment.
Gaussian estimator takes an output switch parameter - contribution of gaussians is no longer always output (saves a lot of time).
Experiment parameter file has data on the number of streams and number of functions to work on.
Fixed memory access issue when combining functions.