Gaussian time delay estimates are automatically normalised. More error functions.
When estimating the time delay with gaussian kernels, the resulting function (output to ~combined.dat) is automatically normalised. Estimating only a single stream with the gaussians still requires normalisation of the returned values.
MSE and RMS error functions added to math util. Also made the corresponding value function shift the data it gets from the gauss vector so that it is all above zero.