Stuttering functions improved, "goodness" of single stream estimates output.
The goodness of single stream estimates is about how good the estimate is when using a stuttered input stream of events.
Modified the estimate function to return void pointers so that data from gaussian estimates can also be returned.
Fixed memory leaks caused by some of the changes, and a few others that were picked up.